Updated spiffeid package scaffolding
Converted the TrustDomain type to a strongly validated struct type with forced
normalization. Provided methods to get a TrustDomain from string and
URI with "Require*" variants that panic instead of returning an error
for convenient use with verifiable constants.

Also renamed some of the ID methods to match trust domain equivalents
for consistency.

Also:
- Fixed up writer-based logger to add newlines
- Fixed up broken examples
- Removed presumed implementation details in the ID type to allow for
implementation flexibility

Signed-off-by: Andrew Harding <azdagron@gmail.com>