Copy bundle content containers in and out of bundles
This change updates the bundles to consistently copy the contents
containers holding the bundle contents (X.509 root slices and JWT key
maps) when those content containers go in and out of the bundle. This is
only a shallow copy of the data structure for the contents, not a deep
copy of the contents themselves.

Specifically, all `From*` methods make a copy of the incoming
containers and the X509Roots and JWTKeys methods return a copy of the
containers.

This prevents accidential aliasing of slice backing arrays and maps.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>