Workload API Sources Fixes and Tests
- Fixes bundle merging behavior in the BundleSource such that a lack of
X.509 or JWT material for a given trust domain causes the relative
bundle getters for X.509 or JWT bundles to return no bundle instead of
an empty bundle
- Reduces code duplication code between the sources to simplify
implementation.
- Adds WaitUntilUpdated methods to each source to allow callers to wait
for changes. An Updated method that returns a channel that can be read
from to detect changes is also added for callers that want to do their
own select{}.
- Adds comprehensive unit-tests for the X509Source, JWTSource, and
BundleSource.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>