Remove non-spec conformant field from Workload proto
The per-SVID federates_with field never made it into the final draft for
federation support in the Workload API. This library takes a dependency
on that field to scope the federated bundles returned by the library to
the workload. SPIRE currently returns this field but will stop at a
future date.

This change removes the field and updates the code to return all
federated bundles.

This change already exists in the v2 version of this library.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>