Replace logging with trace callbacks
The trace callbacks allow for both logging and metric emission by
library consumers.

This change threads trace callbacks into the Workload API client. It
also adjusts the trace callbacks previously added to the tlsconfig
package to follow a uniform pattern that is future proof. Namely:

- start/stop callbacks always receive an info struct, even if empty to
allow for future expansion.
- uniform ordering of parameters/return values

A small helper is added that populates unset trace callback functions.
This removes the burden of sprinkling `if` statements at each callback
call site.

The logging previously supported by the federation handler can be
accomplished by wrapping the bundle source and logging when it is
fetched. The implementation is targeted enough that if it doesn't meet
all the needs of potential consumers, they should be able to write their
own trivially.

Signed-off-by: Andrew Harding <aharding@vmware.com>