Reduce allocations for ID type (#185)
Going from ID to string causes an allocation. This change avoids this
allocation by holding onto the original string and instead tracking
where the path begins inside of that string. Note that this does not
increase memory usage since the backing array of the original string was
already being held by the existing code. This change also reduces the
struct size of the ID type.

Signed-off-by: Andrew Harding <aharding@vmware.com>