Move code out of v2 directory
When this library first went to v2, the old v1 code was kept around so
we opted to put all of the v2 code into a v2 directory. The v1 code
was eventually removed because it was (1) no longer developed, (2) still
available at the old tag, and (3) because its presence was confusing
library consumers.

Now that the v1 code is long gone, there is no reason to have all v2
related materials in a v2 directory.

Signed-off-by: Andrew Harding <azdagron@gmail.com>